# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API for Peminjaman Laboratorium Terpadu Poltekkes Tasikmalaya
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /laboratorium:
    get:
      operationId: getLaboratorium
      tags: [laboratorium]
      summary: Get all laboratories
      parameters:
        - name: jurusanId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of laboratories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Laboratorium"

    post:
      operationId: createLaboratorium
      tags: [laboratorium]
      summary: Create new laboratory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLaboratoriumRequest"
      responses:
        "201":
          description: Laboratory created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Laboratorium"
  /laboratorium/{id}:
    get:
      operationId: getLaboratoriumById
      tags: [laboratorium]
      summary: Get laboratory by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Laboratory data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Laboratorium"

    put:
      operationId: updateLaboratorium
      tags: [laboratorium]
      summary: Update laboratory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLaboratoriumRequest"
      responses:
        "200":
          description: Laboratory updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Laboratorium"

    delete:
      operationId: deleteLaboratorium
      tags: [laboratorium]
      summary: Delete laboratory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Laboratory deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /berita:
    get:
      operationId: getBerita
      tags: [berita]
      summary: Get all news/posts
      parameters:
        - name: kategori
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of news
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Berita"

    post:
      operationId: createBerita
      tags: [berita]
      summary: Create news/post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBeritaRequest"
      responses:
        "201":
          description: News created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Berita"
  /berita/{id}:
    get:
      operationId: getBeritaById
      tags: [berita]
      summary: Get news by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: News data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Berita"

    put:
      operationId: updateBerita
      tags: [berita]
      summary: Update news
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBeritaRequest"
      responses:
        "200":
          description: News updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Berita"

    delete:
      operationId: deleteBerita
      tags: [berita]
      summary: Delete news
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: News deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /galeri:
    get:
      operationId: getGaleri
      tags: [galeri]
      summary: Get all gallery items
      parameters:
        - name: tipe
          in: query
          schema:
            type: string
            enum: [foto, video]
      responses:
        "200":
          description: List of gallery items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Galeri"

    post:
      operationId: createGaleri
      tags: [galeri]
      summary: Create gallery item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGaleriRequest"
      responses:
        "201":
          description: Gallery item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Galeri"
  /galeri/{id}:
    delete:
      operationId: deleteGaleri
      tags: [galeri]
      summary: Delete gallery item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Gallery item deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /dokumen:
    get:
      operationId: getDokumen
      tags: [dokumen]
      summary: Get all lab documents
      parameters:
        - name: laboratoriumId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dokumen"

    post:
      operationId: createDokumen
      tags: [dokumen]
      summary: Upload lab document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDokumenRequest"
      responses:
        "201":
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dokumen"
  /dokumen/{id}:
    delete:
      operationId: deleteDokumen
      tags: [dokumen]
      summary: Delete document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    User:
      type: object
      properties:
        id:
          type: integer
        nama:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, mahasiswa, plp, gudang, dosen]
        nim:
          type: string
          nullable: true
        nip:
          type: string
          nullable: true
        noHp:
          type: string
          nullable: true
        jurusanId:
          type: integer
          nullable: true
        jurusan:
          $ref: "#/components/schemas/Jurusan"
          nullable: true
        laboratoriumId:
          type: integer
          nullable: true
        status:
          type: string
          enum: [menunggu, aktif, nonaktif, ditolak]
        createdAt:
          type: string
          format: date-time
      required: [id, nama, email, role, status, createdAt]
...
    Jurusan:
      type: object
      properties:
        id:
          type: integer
        nama:
          type: string
        kode:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, nama, kode, createdAt]
...
    Laboratorium:
      type: object
      properties:
        id:
          type: integer
        nama:
          type: string
        kode:
          type: string
        lokasi:
          type: string
        kapasitas:
          type: integer
        jurusanId:
          type: integer
          nullable: true
        jurusan:
          $ref: "#/components/schemas/Jurusan"
          nullable: true
        deskripsi:
          type: string
          nullable: true
        fasilitas:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, nama, kode, lokasi, kapasitas, createdAt]
    CreateLaboratoriumRequest:
      type: object
      properties:
        nama:
          type: string
        kode:
          type: string
        lokasi:
          type: string
        kapasitas:
          type: integer
        jurusanId:
          type: integer
          nullable: true
        deskripsi:
          type: string
          nullable: true
        fasilitas:
          type: string
          nullable: true
      required: [nama, kode, lokasi, kapasitas]
...
    Berita:
      type: object
      properties:
        id:
          type: integer
        judul:
          type: string
        konten:
          type: string
        kategori:
          type: string
          nullable: true
        thumbnail:
          type: string
          nullable: true
        penulis:
          $ref: "#/components/schemas/User"
          nullable: true
        penulisId:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, judul, konten, penulisId, createdAt]
    CreateBeritaRequest:
      type: object
      properties:
        judul:
          type: string
        konten:
          type: string
        kategori:
          type: string
          nullable: true
        thumbnail:
          type: string
          nullable: true
      required: [judul, konten]
    Galeri:
      type: object
      properties:
        id:
          type: integer
        judul:
          type: string
        deskripsi:
          type: string
          nullable: true
        tipe:
          type: string
          enum: [foto, video]
        url:
          type: string
        uploaderId:
          type: integer
        uploader:
          $ref: "#/components/schemas/User"
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, judul, tipe, url, uploaderId, createdAt]
    CreateGaleriRequest:
      type: object
      properties:
        judul:
          type: string
        deskripsi:
          type: string
          nullable: true
        tipe:
          type: string
          enum: [foto, video]
        url:
          type: string
      required: [judul, tipe, url]
    Dokumen:
      type: object
      properties:
        id:
          type: integer
        nama:
          type: string
        deskripsi:
          type: string
          nullable: true
        url:
          type: string
        tipe:
          type: string
          nullable: true
        laboratoriumId:
          type: integer
          nullable: true
        laboratorium:
          $ref: "#/components/schemas/Laboratorium"
          nullable: true
        uploaderId:
          type: integer
        uploader:
          $ref: "#/components/schemas/User"
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, nama, url, uploaderId, createdAt]
    CreateDokumenRequest:
      type: object
      properties:
        nama:
          type: string
        deskripsi:
          type: string
          nullable: true
        url:
          type: string
        tipe:
          type: string
          nullable: true
        laboratoriumId:
          type: integer
          nullable: true
      required: [nama, url]
...